First webserver with Bottle
