+++ src/ui/index.html
@@ -50,7 +50,7 @@
             <li><a href="#" class="class_top_button ui-btn ui-icon-delete ui-btn-icon-bottom" onclick="OrderDo('stop');">Stop</a></li>
             <li><a href="#" class="class_top_button ui-btn ui-icon-refresh ui-btn-icon-bottom" onclick="OrderDo('replay');">Replay</a></li>
             <li><a href="#" class="class_top_button ui-btn ui-icon-recycle ui-btn-icon-bottom" onclick="OrderDo('shuffle');">Shuffle</a></li>
-            <li><a href="#" class="class_top_button ui-btn ui-icon-star ui-btn-icon-bottom">Score</a></li>
+            <li><a href="#" class="class_top_button ui-btn ui-icon-star ui-btn-icon-bottom" onclick="OrderDo('score');">Score</a></li>
             <li><a href="#popupDialog" class="class_top_button ui-btn ui-icon-action ui-btn-icon-bottom" data-rel="popup" data-position-to="window" data-transition="pop">Share</a></li>
           </ul>
         </div><!-- /navbar -->
